requirements.txt
langchain==0.3.13
langchain-groq==0.1.0
langchain-mcp-adapters==0.1.12
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
mcp==1.21.0
fastapi
uvicorn
python-dotenv
requests
pydantic
tweepy
